requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
httpx[http2]>=0.27.0
aiolimiter>=1.1.0
tenacity>=9.0.0
pydantic>=2.8.0
pandas>=2.2.2
pyarrow>=17.0.0
structlog>=24.1.0
streamlit>=1.37.0
plotly>=5.23.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
numpy>=2.0.0
